requirements.txt
Python == 3.12.7
python-dotenv==1.0.1
requests==2.32.3
urllib3==1.26.17

